1.文章删除，添加，测试 2.显示图片上传路径

+++ src/main/java/com/student/entity/article/Article.java
@@ -31,7 +31,7 @@ public class Article extends BaseEntity {
     private Long commentId;
 
     /** 分类id */
-    private Long typeId;
+    private Integer typeId;
 
     /** 大图 */
     private String pictureUrl;
